+++ pkg/app/mutate_app.go
@@ -9,7 +9,7 @@ import (
 	"github.com/giantswarm/apiextensions/v3/pkg/apis/application/v1alpha1"
 	"github.com/giantswarm/apiextensions/v3/pkg/clientset/versioned"
 	"github.com/giantswarm/apiextensions/v3/pkg/label"
-	"github.com/giantswarm/app/v3/pkg/key"
+	"github.com/giantswarm/app/v4/pkg/key"
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
@@ -113,7 +113,10 @@ func (m *Mutator) MutateApp(ctx context.Context, app v1alpha1.App) ([]mutator.Pa
 		return nil, nil
 	}
 
-	if ver.Major() < 3 {
+	// If the app CR does not have the unique version and is < 3.0.0 we skip
+	// the defaulting logic. This is so the admission controller is not enabled
+	// for existing platform releases.
+	if key.VersionLabel(app) != uniqueAppCRVersion && ver.Major() < 3 {
 		m.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("skipping mutation of app %#q in namespace %#q due to version label %#q", app.Name, app.Namespace, appVersionLabel))
 		return nil, nil
 	}